requirements.txt
numpy==1.21.5
websockets==15.0.1
PyYAML
requests